# src/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Vehicle Service API",
    "description": "API for accessing vehicle information, articles, maintenance schedules, parts, and related content",
    "version": "1.0.0",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "https://us-central1-vehapi-torque.cloudfunctions.net/motorApiAuthProxy",
      "description": "Production server (Motor Proxy)"
    }
  ],
...
  "paths": {
...
    "/auth/start": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Start authentication",
        "description": "Manually trigger authentication. Authentication also starts automatically when a 401 response is received. Use this endpoint to pre-authenticate or retry failed authentication.",
        "operationId": "startAuth",
        "responses": {
          "200": {
            "description": "Authentication started",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": ["started"],
                      "description": "Status indicating authentication has started"
                    },
                    "message": {
                      "type": "string",
                      "description": "Instructions for polling authentication status"
                    }
                  },
                  "required": ["status", "message"]
                },
                "example": {
                  "status": "started",
                  "message": "Authentication started. Poll /auth/status for progress."
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
  },
  "components": {
...
    "responses": {
...
      "InternalServerError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Internal Server Error"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  }
}